\name{RdfTerm-class}
\Rdversion{1.1}
\docType{class}
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
% Alias
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
\alias{RdfTerm-class}
%
\title{Class \code{"RdfTerm"}}
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
% Description
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
\description{RdfTerm class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
    \describe{
        \item{\code{prefix}:}{Object of class \code{"character"} ~~ }
    }
}
\section{Methods}{
No methods defined with class "RdfTerm" in the signature.
}
\author{Wolfgang Kaisers}
\examples{
showClass("RdfTerm")
}
\keyword{classes}
